# workspace/fairsplit/api-spec.yaml
openapi: 3.1.0
info:
  title: FairSplit API
  version: 1.0.0
  description: |
    Zero-dependency shared expense tracker API.

    ## Authentication
    All group-scoped endpoints require a valid member session token.
    The token is issued by `POST /api/v1/groups` (group creation) or
    `POST /api/v1/groups/{group_id}/members` (join via invite link).

    **Primary:** `httpOnly` cookie named `fairsplit_member_token`.
    **Fallback:** `Authorization: Bearer <token>` header (API clients).

    The token is a JWT signed with HS256 containing:
    ```json
    {
      "sub": "<member_id>",
      "group_id": "<group_id>",
      "is_admin": true,
      "iat": 1713200000,
      "exp": 1715792000
    }
    ```

    ## Response Envelopes
    **Singleton responses** (auth, single-resource GET/POST/PATCH/DELETE):
    The resource object is returned directly at the top level. No wrapper.

    **List responses** (collection GET endpoints):
    ```json
    {
      "data": [...],
      "total": 23,
      "page": 1,
      "per_page": 20
    }
    ```

    **Error responses** (all 4xx and 5xx):
    ```json
    {
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Human-readable description.",
        "field": "amount",
        "request_id": "req_abc123"
      }
    }
    ```

    ## Idempotency
    Mutation endpoints (`POST` for create, `DELETE` for soft-delete) accept an
    `X-Idempotency-Key` header containing a client-generated UUID. The server
...

  contact:
...
servers:
...
paths:
...
  /groups/{group_id}/members:
    get:
      operationId: listMembers
      summary: List all members in a group
      description: |
        Returns all members of the group sorted alphabetically by display name.
        Requires a valid member session token for the specified group.
      tags: [members]
      security:
        - MemberCookie: []
        - BearerToken: []
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: object
                required: [data, total, page, per_page]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Member"
                  total:
                    type: integer
                    example: 6
                  page:
                    type: integer
                    example: 1
                  per_page:
                    type: integer
                    example: 100
              example:
                data:
                  - id: "a3bb1891-1f4e-4e7a-8c4f-6d4a1e2b3c4d"
                    display_name: "Maya"
                    is_admin: true
                    joined_at: "2026-04-15T14:00:00Z"
                  - id: "b4cc2902-2g5f-5f8b-9d5g-7e5b2f3c4d5e"
                    display_name: "Daniel"
                    is_admin: false
                    joined_at: "2026-04-15T14:05:00Z"
                total: 6
                page: 1
                per_page: 100
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /join/{invite_token}:
    post:
      operationId: joinGroup
      summary: Join a group via invite link
      description: |
        Join a group by providing a display name. No email or password required.
        Returns a session token in the response body AND sets an httpOnly cookie.
        The `invite_token` is the UUID from the shareable link:
        `{base_url}/join/{invite_token}`.

        If `invite_token` does not match any active group, returns 404 with
        `INVALID_INVITE_TOKEN`.

        If the group is archived, returns 409 with `GROUP_ARCHIVED`.

        **Duplicate display names:** Allowed. If the submitted name already exists
        in the group, the server appends " (2)" (or higher) and returns the
        modified name in the response. The client should display this modified name.

        **Rate limit:** 30 requests per IP per hour.

        Idempotent: if an `X-Idempotency-Key` is provided and the client
        accidentally submits twice (e.g., double-tap), the second request returns
        the original 201 without creating a second member record.
      tags: [members]
      parameters:
        - name: invite_token
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The group invite token from the shareable link URL
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinGroupRequest"
            example:
              display_name: "Daniel"
      responses:
        "201":
          description: Joined group successfully
          headers:
            Set-Cookie:
              description: httpOnly session cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinGroupResponse"
              example:
                group:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  name: "Nashville Trip"
                  status: active
                  default_currency: USD
                  member_count: 7
                member:
                  id: "b4cc2902-2g5f-5f8b-9d5g-7e5b2f3c4d5e"
                  display_name: "Daniel"
                  is_admin: false
                  joined_at: "2026-04-15T14:05:00Z"
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        "404":
          description: Invite token not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: INVALID_INVITE_TOKEN
                  message: "This link is no longer valid. Ask the group admin for a new one."
                  request_id: req_abc123
        "409":
          description: Group is archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: GROUP_ARCHIVED
                  message: "This group has been closed. You can view it in read-only mode."
                  request_id: req_abc123
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
  /groups/{group_id}/members/me:
    get:
      operationId: getCurrentMember
      summary: Get the current authenticated member's profile
      description: |
        Returns the member record for the currently authenticated session.
        Used on page load to confirm session validity and get the current
        member's ID and display name.
      tags: [members]
      security:
        - MemberCookie: []
        - BearerToken: []
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: Current member profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
              example:
                id: "a3bb1891-1f4e-4e7a-8c4f-6d4a1e2b3c4d"
                display_name: "Maya"
                is_admin: true
                joined_at: "2026-04-15T14:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ============================================================
  # EXPENSES
  # ============================================================
...
  /groups/{group_id}/settle-up:
    get:
      operationId: getSettleUp
      summary: Compute the minimum-transfer settle-up plan
      description: |
        Runs the greedy max-heap minimum-transfer algorithm on the current
        net balances and returns an ordered list of transfers that, if
        completed, would bring all members' net balances to zero.

        **Algorithm:** O(n log n) greedy max-heap (see technical-spec.md §1
        for full pseudocode and correctness guarantees).

        **Correctness note:** The algorithm is a heuristic. It produces the
        exact minimum number of transfers for groups up to ~20 members with
        all tested balance configurations. For larger groups, it produces a
        near-optimal result. This is documented in the product FAQ.

        If all balances are zero, returns an empty `transfers` array with
        `all_settled: true`.

        This endpoint is never cached — it always reflects the current state
        of expenses and settlements.

        **Performance:** < 100ms p95 for groups up to 50 members.
      tags: [settle-up]
      security:
        - MemberCookie: []
        - BearerToken: []
      parameters:
        - $ref: "#/components/parameters/GroupId"
        - name: currency
          in: query
          required: false
          schema:
            type: string
            pattern: "^[A-Z]{3}$"
            default: "USD"
          description: |
            ISO 4217 currency code. For multi-currency groups, the settle-up
            is computed independently per currency. This parameter selects
            which currency to compute. Defaults to the group's default_currency.
      responses:
        "200":
          description: Settle-up plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettleUpResponse"
              examples:
                with_transfers:
                  summary: Group with outstanding balances
                  value:
                    group_name: "Nashville Trip"
                    currency: USD
                    all_settled: false
                    transfer_count: 2
                    transfers:
                      - debtor_id: "b4cc2902-2g5f-5f8b-9d5g-7e5b2f3c4d5e"
                        debtor_name: "Daniel"
                        creditor_id: "a3bb1891-1f4e-4e7a-8c4f-6d4a1e2b3c4d"
                        creditor_name: "Maya"
                        amount_cents: 4200
                        amount_display: "$42.00"
                        currency: USD
                      - debtor_id: "d6ee4124-4i7h-7h0d-1f7i-9g7d4h5e6f7g"
                        debtor_name: "Jake"
                        creditor_id: "e7ff5235-5j8i-8i1e-2g8j-0h8e5i6f7g8h"
                        creditor_name: "Tom"
                        amount_cents: 1850
                        amount_display: "$18.50"
                        currency: USD
                    clipboard_text: "FairSplit — Nashville Trip Settle-Up\n• Daniel pays Maya $42.00\n• Jake pays Tom $18.50\nPowered by FairSplit: https://fairsplit.app/groups/550e8400-e29b-41d4-a716-446655440000"
                    computed_at: "2026-04-15T18:30:00Z"
                all_settled:
                  summary: All balances are zero
                  value:
                    group_name: "Nashville Trip"
                    currency: USD
                    all_settled: true
                    transfer_count: 0
                    transfers: []
                    clipboard_text: "FairSplit — Nashville Trip\nEveryone is settled up. Nothing to pay."
                    computed_at: "2026-04-15T18:30:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ============================================================
  # SETTLEMENTS
  # ============================================================
...
components:
  securitySchemes:
    MemberCookie:
      type: apiKey
      in: cookie
      name: fairsplit_member_token
      description: |
        httpOnly session cookie issued on group creation or group join.
        Set by the server. Cannot be accessed by JavaScript.
        Expiry: 30 days, rolling (reset on each valid request).
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Fallback for API clients that cannot use cookies.
        Same JWT as the cookie value. Pass in Authorization: Bearer <token> header.
  parameters:
    GroupId:
      name: group_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: The group's UUID
...
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: |
        Client-generated UUID for idempotency protection. Generate one UUID
        when the form opens and reuse it on all retries. The server returns
        the original response for duplicate keys (within 24 hours).
  responses:
    Unauthorized:
      description: No session or session expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            no_session:
              value:
                error:
                  code: NO_SESSION
                  message: "No session token found. Please join the group first."
                  request_id: req_abc123
            session_expired:
              value:
                error:
                  code: SESSION_EXPIRED
                  message: "Your session has expired. Please re-enter your name to continue."
                  request_id: req_abc123
    Forbidden:
      description: Valid session but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            permission_denied:
              value:
                error:
                  code: PERMISSION_DENIED
                  message: "You can only edit expenses you logged, or you must be the group admin."
                  request_id: req_abc123
            wrong_group:
              value:
                error:
                  code: MEMBER_WRONG_GROUP
                  message: "Your session token is for a different group."
                  request_id: req_abc123
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: GROUP_NOT_FOUND
              message: "Group not found. The link may be invalid."
              request_id: req_abc123
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_ERROR
              message: "Amount must be greater than zero."
              field: amount_cents
              request_id: req_abc123
    RateLimitExceeded:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: RATE_LIMIT_EXCEEDED
              message: "Too many requests. Please wait 42 seconds."
              request_id: req_abc123
...
  schemas:
...
    Member:
      type: object
      required: [id, display_name, is_admin, joined_at]
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
          minLength: 1
          maxLength: 40
        is_admin:
          type: boolean
        joined_at:
          type: string
          format: date-time
...
    Transfer:
      type: object
      required:
        - debtor_id
        - debtor_name
        - creditor_id
        - creditor_name
        - amount_cents
        - amount_display
        - currency
      properties:
        debtor_id:
          type: string
          format: uuid
          description: The member who owes money (will pay)
        debtor_name:
          type: string
        creditor_id:
          type: string
          format: uuid
          description: The member who is owed money (will receive)
        creditor_name:
          type: string
        amount_cents:
          type: integer
          minimum: 1
          description: Transfer amount in integer cents
        amount_display:
          type: string
          description: Formatted display string (e.g., "$42.00")
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
    SettleUpResponse:
      type: object
      required:
        - group_name
        - currency
        - all_settled
        - transfer_count
        - transfers
        - clipboard_text
        - computed_at
      properties:
        group_name:
          type: string
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
        all_settled:
          type: boolean
          description: True if all net balances are zero — no transfers needed
        transfer_count:
          type: integer
          minimum: 0
        transfers:
          type: array
          items:
            $ref: "#/components/schemas/Transfer"
          description: |
            Ordered list of transfers. Empty if all_settled is true.
            Order: largest transfer first (by amount_cents DESC).
        clipboard_text:
          type: string
          description: |
            Pre-formatted plain text for copying to a group chat.
            Format:
              "FairSplit — {group_name} Settle-Up\n
               • {debtor} pays {creditor} {amount_display} {currency}\n
               ...\n
               Powered by FairSplit: {group_url}"
        computed_at:
          type: string
          format: date-time
...
    JoinGroupRequest:
      type: object
      required: [display_name]
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 40
          description: |
            Display name for the joining member.
            If this name already exists in the group, the server appends " (2)",
            " (3)", etc. The modified name is returned in the response.
    JoinGroupResponse:
      type: object
      required: [group, member, token]
      properties:
        group:
          type: object
          required: [id, name, status, default_currency, member_count]
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            status:
              type: string
              enum: [active, archived]
            default_currency:
              type: string
            member_count:
              type: integer
        member:
          $ref: "#/components/schemas/Member"
        token:
          type: string
          description: Signed JWT. Also set as an httpOnly cookie.
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - GROUP_NOT_FOUND
                - MEMBER_NOT_FOUND
                - EXPENSE_NOT_FOUND
                - SETTLEMENT_NOT_FOUND
                - INVALID_INVITE_TOKEN
                - GROUP_ARCHIVED
                - GROUP_HAS_OUTSTANDING_BALANCES
                - PERMISSION_DENIED
                - NO_SESSION
                - SESSION_EXPIRED
                - SESSION_INVALID
                - MEMBER_WRONG_GROUP
                - IDEMPOTENCY_CONFLICT
                - SPLIT_SUM_MISMATCH
                - RATE_LIMIT_EXCEEDED
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable error description
            field:
              type: string
              nullable: true
              description: The specific field that failed validation (for VALIDATION_ERROR only)
            request_id:
              type: string
              description: Unique request identifier for log correlation